drop support for markdown in item bodies
